Refactor interactive executor classes
