+++ tests/test_local_executor.py
@@ -9,11 +9,9 @@ import numpy as np
 
 from executorlib.base.executor import ExecutorBase
 from executorlib.standalone.interactive.spawner import MpiExecSpawner
-from executorlib.interactive.shared import (
-    InteractiveExecutor,
-    InteractiveStepExecutor,
-    execute_parallel_tasks,
-)
+from executorlib.interactive.shared import execute_tasks
+from executorlib.interactive.blockallocation import BlockAllocationExecutor
+from executorlib.interactive.onetoone import OneTaskPerProcessExecutor
 from executorlib.standalone.interactive.backend import call_funct
 from executorlib.standalone.serialize import cloudpickle_register
 
@@ -66,7 +64,7 @@ def sleep_one(i):
 
 class TestPyMpiExecutorSerial(unittest.TestCase):
     def test_pympiexecutor_two_workers(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=2,
             executor_kwargs={},
             spawner=MpiExecSpawner,
@@ -80,7 +78,7 @@ class TestPyMpiExecutorSerial(unittest.TestCase):
             self.assertTrue(fs_2.done())
 
     def test_pympiexecutor_one_worker(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={},
             spawner=MpiExecSpawner,
@@ -96,7 +94,7 @@ class TestPyMpiExecutorSerial(unittest.TestCase):
 
 class TestPyMpiExecutorStepSerial(unittest.TestCase):
     def test_pympiexecutor_two_workers(self):
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=2,
             executor_kwargs={},
             spawner=MpiExecSpawner,
@@ -110,7 +108,7 @@ class TestPyMpiExecutorStepSerial(unittest.TestCase):
             self.assertTrue(fs_2.done())
 
     def test_pympiexecutor_one_worker(self):
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=1,
             executor_kwargs={},
             spawner=MpiExecSpawner,
@@ -129,7 +127,7 @@ class TestPyMpiExecutorStepSerial(unittest.TestCase):
 )
 class TestPyMpiExecutorMPI(unittest.TestCase):
     def test_pympiexecutor_one_worker_with_mpi(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -140,7 +138,7 @@ class TestPyMpiExecutorMPI(unittest.TestCase):
             self.assertTrue(fs_1.done())
 
     def test_pympiexecutor_one_worker_with_mpi_multiple_submissions(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -160,7 +158,7 @@ class TestPyMpiExecutorMPI(unittest.TestCase):
         )
 
     def test_pympiexecutor_one_worker_with_mpi_echo(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -175,7 +173,7 @@ class TestPyMpiExecutorMPI(unittest.TestCase):
 )
 class TestPyMpiStepExecutorMPI(unittest.TestCase):
     def test_pympiexecutor_one_worker_with_mpi(self):
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=2,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -186,7 +184,7 @@ class TestPyMpiStepExecutorMPI(unittest.TestCase):
             self.assertTrue(fs_1.done())
 
     def test_pympiexecutor_one_worker_with_mpi_multiple_submissions(self):
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=2,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -206,7 +204,7 @@ class TestPyMpiStepExecutorMPI(unittest.TestCase):
         )
 
     def test_pympiexecutor_one_worker_with_mpi_echo(self):
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=2,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -218,7 +216,7 @@ class TestPyMpiStepExecutorMPI(unittest.TestCase):
 
 class TestPyMpiExecutorInitFunction(unittest.TestCase):
     def test_internal_memory(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={
                 "cores": 1,
@@ -246,7 +244,7 @@ class TestPyMpiExecutorInitFunction(unittest.TestCase):
         q.put({"fn": get_global, "args": (), "kwargs": {}, "future": f})
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
-        execute_parallel_tasks(
+        execute_tasks(
             future_queue=q,
             cores=1,
             openmpi_oversubscribe=False,
@@ -259,7 +257,7 @@ class TestPyMpiExecutorInitFunction(unittest.TestCase):
 
 class TestFuturePool(unittest.TestCase):
     def test_pool_serial(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 1},
             spawner=MpiExecSpawner,
@@ -274,7 +272,7 @@ class TestFuturePool(unittest.TestCase):
         self.assertEqual(output.result(), np.array(4))
 
     def test_executor_multi_submission(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 1},
             spawner=MpiExecSpawner,
@@ -287,7 +285,7 @@ class TestFuturePool(unittest.TestCase):
             self.assertTrue(fs_2.done())
 
     def test_shutdown(self):
-        p = InteractiveExecutor(
+        p = BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 1},
             spawner=MpiExecSpawner,
@@ -303,7 +301,7 @@ class TestFuturePool(unittest.TestCase):
             fs2.result()
 
     def test_pool_serial_map(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 1},
             spawner=MpiExecSpawner,
@@ -313,7 +311,7 @@ class TestFuturePool(unittest.TestCase):
 
     def test_executor_exception(self):
         with self.assertRaises(RuntimeError):
-            with InteractiveExecutor(
+            with BlockAllocationExecutor(
                 max_workers=1,
                 executor_kwargs={"cores": 1},
                 spawner=MpiExecSpawner,
@@ -323,7 +321,7 @@ class TestFuturePool(unittest.TestCase):
 
     def test_executor_exception_future(self):
         with self.assertRaises(RuntimeError):
-            with InteractiveExecutor(
+            with BlockAllocationExecutor(
                 max_workers=1,
                 executor_kwargs={"cores": 1},
                 spawner=MpiExecSpawner,
@@ -344,7 +342,7 @@ class TestFuturePool(unittest.TestCase):
             "openmpi_oversubscribe": False,
             "max_workers": 1,
         }
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={
                 "cores": 2,
@@ -372,7 +370,7 @@ class TestFuturePool(unittest.TestCase):
             "openmpi_oversubscribe": False,
             "max_cores": 2,
         }
-        with InteractiveStepExecutor(
+        with OneTaskPerProcessExecutor(
             max_cores=2,
             executor_kwargs={
                 "cores": 2,
@@ -392,7 +390,7 @@ class TestFuturePool(unittest.TestCase):
         skip_mpi4py_test, "mpi4py is not installed, so the mpi4py tests are skipped."
     )
     def test_pool_multi_core(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -410,7 +408,7 @@ class TestFuturePool(unittest.TestCase):
         skip_mpi4py_test, "mpi4py is not installed, so the mpi4py tests are skipped."
     )
     def test_pool_multi_core_map(self):
-        with InteractiveExecutor(
+        with BlockAllocationExecutor(
             max_workers=1,
             executor_kwargs={"cores": 2},
             spawner=MpiExecSpawner,
@@ -428,7 +426,7 @@ class TestFuturePool(unittest.TestCase):
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
         with self.assertRaises(TypeError):
-            execute_parallel_tasks(
+            execute_tasks(
                 future_queue=q,
                 cores=1,
                 openmpi_oversubscribe=False,
@@ -444,7 +442,7 @@ class TestFuturePool(unittest.TestCase):
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
         with self.assertRaises(TypeError):
-            execute_parallel_tasks(
+            execute_tasks(
                 future_queue=q,
                 cores=1,
                 openmpi_oversubscribe=False,
@@ -459,7 +457,7 @@ class TestFuturePool(unittest.TestCase):
         q.put({"fn": calc_array, "args": (), "kwargs": {"i": 2}, "future": f})
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
-        execute_parallel_tasks(
+        execute_tasks(
             future_queue=q,
             cores=1,
             openmpi_oversubscribe=False,
@@ -477,7 +475,7 @@ class TestFuturePool(unittest.TestCase):
         q.put({"fn": calc_array, "args": (), "kwargs": {"i": 2}, "future": f})
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
-        execute_parallel_tasks(
+        execute_tasks(
             future_queue=q,
             cores=2,
             openmpi_oversubscribe=False,
@@ -500,7 +498,7 @@ class TestFuturePoolCache(unittest.TestCase):
         q.put({"fn": calc, "args": (), "kwargs": {"i": 1}, "future": f})
         q.put({"shutdown": True, "wait": True})
         cloudpickle_register(ind=1)
-        execute_parallel_tasks(
+        execute_tasks(
             future_queue=q,
             cores=1,
             openmpi_oversubscribe=False,
@@ -519,7 +517,7 @@ class TestFuturePoolCache(unittest.TestCase):
         q.put({"fn": calc_array, "args": (), "kwargs": {}, "future": f})
         cloudpickle_register(ind=1)
         with self.assertRaises(TypeError):
-            execute_parallel_tasks(
+            execute_tasks(
                 future_queue=q,
                 cores=1,
                 openmpi_oversubscribe=False,
